requirements.txt
bitstring
Pillow
